# Optimise decimal casting for infallible conversions