--- a/arrow-cast/src/cast/decimal.rs
+++ b/arrow-cast/src/cast/decimal.rs
@@ -88,33 +88,18 @@ where
     T: DecimalType,
     T::Native: DecimalCast + ArrowNativeTypeOp,
 {
-    let error = cast_decimal_to_decimal_error::<T, T>(output_precision, output_scale);
     let delta_scale = input_scale - output_scale;
-    let div = T::Native::from_decimal(10_i128)
-        .unwrap()
-        .pow_checked((input_scale - output_scale) as u32)?;
-
-    let half = div.div_wrapping(T::Native::from_usize(2).unwrap());
-    let half_neg = half.neg_wrapping();
-
-    let f = |x: T::Native| {
-        // div is >= 10 and so this cannot overflow
-        let d = x.div_wrapping(div);
-        let r = x.mod_wrapping(div);
-
-        // Round result
-        let adjusted = match x >= T::Native::ZERO {
-            true if r >= half => d.add_wrapping(T::Native::ONE),
-            false if r <= half_neg => d.sub_wrapping(T::Native::ONE),
-            _ => d,
-        };
-        T::Native::from_decimal(adjusted)
-    };
     // if the reudction of the input number through scaling (dividing) is greater
     // than a possible precision loss (plus potential increase via rounding) every input number will fit into the output type
     let is_infallible_cast = (input_precision as i8) - delta_scale < (output_precision as i8);
 
-    Ok(if is_infallible_cast {
+    if is_infallible_cast {
+        let div = T::Native::from_decimal(10_i128)
+            .unwrap()
+            .pow_checked(delta_scale as u32)?;
+
+        let half = div.div_wrapping(T::Native::from_usize(2).unwrap());
+        let half_neg = half.neg_wrapping();
         let f = |x: T::Native| {
             // div is >= 10 and so this cannot overflow
             let d = x.div_wrapping(div);
@@ -127,15 +112,16 @@ where
                 _ => d,
             }
         };
-        array.unary(f)
-    } else if cast_options.safe {
-        array.unary_opt(|x| f(x).filter(|v| T::is_valid_decimal_precision(*v, output_precision)))
+        Ok(array.unary(f))
     } else {
-        array.try_unary(|x| {
-            f(x).ok_or_else(|| error(x))
-                .and_then(|v| T::validate_decimal_precision(v, output_precision).map(|_| v))
-        })?
-    })
+        convert_to_smaller_scale_decimal::<T, T>(
+            array,
+            input_scale,
+            output_precision,
+            output_scale,
+            cast_options,
+        )
+    }
 }
 
 pub(crate) fn convert_to_smaller_scale_decimal<I, O>(
@@ -195,29 +181,26 @@ where
     T: DecimalType,
     T::Native: DecimalCast + ArrowNativeTypeOp,
 {
-    let error = cast_decimal_to_decimal_error::<T, T>(output_precision, output_scale);
-
     let delta_scale = output_scale - input_scale;
-    let mul = T::Native::from_decimal(10_i128)
-        .unwrap()
-        .pow_checked((output_scale - input_scale) as u32)?;
-
-    let f = |x| T::Native::from_decimal(x).and_then(|x| x.mul_checked(mul).ok());
     // if the gain in precision (digits) is greater than the multiplication due to scaling
     // every number will fit into the output type
     let is_infallible_cast = (input_precision as i8) + delta_scale <= (output_precision as i8);
 
-    Ok(if is_infallible_cast {
+    if is_infallible_cast {
+        let mul = T::Native::from_decimal(10_i128)
+            .unwrap()
+            .pow_checked(delta_scale as u32)?;
         let f = |x: T::Native| x.mul_wrapping(mul);
-        array.unary(f)
-    } else if cast_options.safe {
-        array.unary_opt(|x| f(x).filter(|v| T::is_valid_decimal_precision(*v, output_precision)))
+        Ok(array.unary(f))
     } else {
-        array.try_unary(|x| {
-            f(x).ok_or_else(|| error(x))
-                .and_then(|v| T::validate_decimal_precision(v, output_precision).map(|_| v))
-        })?
-    })
+        convert_to_bigger_or_equal_scale_decimal::<T, T>(
+            array,
+            input_scale,
+            output_precision,
+            output_scale,
+            cast_options,
+        )
+    }
 }
 
 pub(crate) fn convert_to_bigger_or_equal_scale_decimal<I, O>(