@alamb 2025-02-07
FYI @andygrove @parthchandra and @himadripal who I think currently care quite a bit about Decimals