@himadripal 2025-02-07
thanks @aweltsch for the contribution, there are many existing tests for decimal conversion,  should those be ported over to using this format. (i.e using DecimalCastTestConfig)