@aweltsch 2025-02-09
> > thanks @aweltsch for the contribution, there are many existing tests for decimal conversion, should those be ported over to using this format. (i.e using DecimalCastTestConfig)?
> 
> When introducing the struct I didn't see the necessity for the other test cases, because they only deal with one combination for input/output precision & scale. Since I've introduced it, I'm happy to convert the other cases to have a more consistent code base. Though I would prefer to do it in a different PR TBH.

After having put the conversion logic in a central place it now applies also to conversions for Decimal256 -> Decimal128 (or the other way around). Hence I want to increase the test coverage a little bit more, but I'm still trying to figure out what a simple yet flexible solution would look like. Not sure if the result will be applicable to the rest of the test cases.