@aweltsch 2025-02-20
Thanks @himadripal, @parthchandra & @tustvold for taking a look. Sorry that clippy failed, I had the pre-commit hook enabled, but for some reason clippy doesn't run with `--workspace --all-targets --all-features -- -D warnings` there. Does anybody know if there's a particular reason the hook runs a different configuration than is recommended in the CONTRIBUTING.md?